feat: Phase 3 — Benchmarks, validation, and critical fixes
Benchmarks (Qwen2.5-7B-Instruct-4bit on M1 Max):
- Quality: 4-bit cosine sim 0.95, 2-bit 0.77 vs FP16 baseline
- Memory: 4.4x compression at 4K context (K3/V2)
- Speed: ~47% decode overhead (software dequant, no Metal kernel)

Critical fixes:
- Float32 norms in cache (float16 overflows for Qwen's extreme key norms)
- Float32 norm computation in quantizer (prevents intermediate overflow)
- Auto-detect outlier layers (Qwen layers 0,1,3,27 kept in FP16)

Benchmark scripts: bench_quality.py, bench_memory.py, bench_speed.py,
needle_haystack.py

108 tests pass.

+++ tests/test_phase1.py
@@ -354,7 +354,7 @@ class TestTurboQuantMSE:
         qt = tq.quantize(x)
         assert qt.bits == 4
         assert qt.d == 128
-        assert qt.norms.dtype == mx.float16
+        assert qt.norms.dtype == mx.float32
         assert qt.qjl_signs is None
